Add files via upload
The goal of the study is to analyse the data as precisely as possible to leverage insights which can save lives. 
In order to do this we have created a promising logistic regression algorithm that predicts which factors influence the mortality rate.
The data collected involves people who have been vaccinated and includes medical records, incidences throughout their vaccination process such as place of administration, how it was administered and if they were hospitalized.